Add A-Varargs.
Refactored to follow better OOP practice.
