+++ src/main/java/duke/command/CommandUnknown.java
@@ -1,0 +1,37 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+/**
+ * Handles when Duke do not understand the command given.
+ */
+public class CommandUnknown extends Command {
+    /**
+     * Constructor for the UnknownCommand.
+     */
+    public CommandUnknown() {}
+
+    /**
+     * Informs user that command given is not understood.
+     *
+     * @param taskList The TaskList that is saved in Duke.
+     * @param ui       The Ui of Duke.
+     * @param storage  The Storage of Duke.
+     * @return String from UI.
+     */
+    @Override
+    public String execute(TaskList taskList, Ui ui, Storage storage) {
+        return ui.displayUnknownUi();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == this) {
+            return true;
+        }
+
+        return obj instanceof CommandUnknown;
+    }
+}
