+++ src/main/java/duke/ui/Ui.java
@@ -1,7 +1,11 @@
-package duke;
+package duke.ui;
 
 import java.util.ArrayList;
 import java.util.Scanner;
+
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
 
 /**
  * Deals with interactions with the user.
@@ -22,8 +26,9 @@ public class Ui {
      *
      * @param taskList The TaskList.
      * @param task The Task that has been added.
+     * @return String representation of addUi.
      */
-    public String displayAddUi(TaskList taskList, TaskList.Task task) {
+    public String displayAddUi(TaskList taskList, Task task) {
         return "Got it. I've added this task:\n" + "  " + task + "\n"
                 + "Now you have " + taskList.size() + " tasks in the list.";
     }
@@ -32,8 +37,9 @@ public class Ui {
      * Displays UI when a task has been marked as done.
      *
      * @param task The Task that has been marked as done.
+     * @return String representation of doneUi.
      */
-    public String displayDoneUi(TaskList.Task task) {
+    public String displayDoneUi(Task task) {
         if (task != null) {
             return "Nice! I've marked this task as done:\n  " + task;
         } else {
@@ -46,8 +52,9 @@ public class Ui {
      *
      * @param taskList The TaskList.
      * @param task The Task that has been deleted.
+     * @return String representation of deleteUi.
      */
-    public String displayDeleteUi(TaskList taskList, TaskList.Task task) {
+    public String displayDeleteUi(TaskList taskList, Task task) {
         return "Noted. I've removed this task:\n  " + task + "\nNow you have "
                 + taskList.size() + " tasks in the list.";
     }
@@ -56,6 +63,7 @@ public class Ui {
      * Displays UI when a command to display TaskList has been given.
      *
      * @param taskList The TaskList.
+     * @return String representation of listUi.
      */
     public String displayListUi(TaskList taskList) {
         if (taskList.isEmpty()) {
@@ -67,6 +75,8 @@ public class Ui {
 
     /**
      * Displays UI when a command to exit Duke has been given.
+     *
+     * @return String representation of exitUi.
      */
     public String displayExitUi() {
         return "Bye. Hope to see you again soon!";
@@ -74,17 +84,20 @@ public class Ui {
 
     /**
      * Displays UI when an unknown command has been given.
+     *
+     * @return String representation of unknownUi.
      */
     public String displayUnknownUi() {
-        return "☹ OOPS!!! I'm sorry, but I don't know what that means :-(";
+        return "OOPS!!! I'm sorry, but I don't know what that means :-(";
     }
 
     /**
      * Displays UI when a command to find a keyword has been given.
      *
      * @param relatedList The list of related tasks found in TaskList.
+     * @return String representation of findUi.
      */
-    public String displayFindUi(ArrayList<TaskList.Task> relatedList) {
+    public String displayFindUi(ArrayList<Task> relatedList) {
         if (relatedList.isEmpty()) {
             return "Unfortunately, there's no task that matches your keyword :-(";
         } else {
@@ -103,14 +116,30 @@ public class Ui {
 
     }
 
+    /**
+     * Displays UI when a `DukeException` is thrown.
+     *
+     * @param e The DukeException thrown.
+     * @return String representation of the `DukeException`.
+     */
     public String displayDukeExceptionMessage(DukeException e) {
         return e.getMessage();
     }
 
+    /**
+     * Greets the user.
+     *
+     * @return A simple greeting.
+     */
     public String greet() {
         return "Hello! I'm Duke\nWhat can I do for you?";
     }
 
+    /**
+     * Greets existing user.
+     *
+     * @return A greeting for existing user.
+     */
     public String greetExistingUser() {
         return "Welcome back! :-)\nWhat can I do for you?";
     }
