+++ src/test/java/duke/TaskListTest.java
@@ -5,6 +5,9 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import org.junit.jupiter.api.Test;
 
+import duke.exception.DukeException;
+import duke.task.TaskList;
+
 public class TaskListTest {
     @Test
     public void taskListConstructor_emptyString_dukeExceptionThrown() {
